@@ -13,6 +13,7 @@ namespace App\Command;
 
 use App\Entity\User;
 use App\Repository\UserRepository;
+use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
@@ -37,22 +38,19 @@ use Symfony\Component\Mime\Email;
  *
  * @author Javier Eguiluz <javier.eguiluz@gmail.com>
  */
+#[AsCommand(
+    name: 'app:list-users',
+    description: 'Lists all the existing users',
+    aliases: ['app:users']
+)]
 class ListUsersCommand extends Command
 {
-    // a good practice is to use the 'app:' prefix to group all your custom application commands
-    protected static $defaultName = 'app:list-users';
-
-    private $mailer;
-    private $emailSender;
-    private $users;
-
-    public function __construct(MailerInterface $mailer, string $emailSender, UserRepository $users)
-    {
+    public function __construct(
+        private MailerInterface $mailer,
+        private string $emailSender,
+        private UserRepository $users
+    ) {
         parent::__construct();
-
-        $this->mailer = $mailer;
-        $this->emailSender = $emailSender;
-        $this->users = $users;
     }
 
     /**
@@ -61,23 +59,21 @@ class ListUsersCommand extends Command
     protected function configure(): void
     {
         $this
-            ->setDescription('Lists all the existing users')
             ->setHelp(<<<'HELP'
-The <info>%command.name%</info> command lists all the users registered in the application:
+                The <info>%command.name%</info> command lists all the users registered in the application:
 
-  <info>php %command.full_name%</info>
+                  <info>php %command.full_name%</info>
 
-By default the command only displays the 50 most recent users. Set the number of
-results to display with the <comment>--max-results</comment> option:
+                By default the command only displays the 50 most recent users. Set the number of
+                results to display with the <comment>--max-results</comment> option:
 
-  <info>php %command.full_name%</info> <comment>--max-results=2000</comment>
+                  <info>php %command.full_name%</info> <comment>--max-results=2000</comment>
 
-In addition to displaying the user list, you can also send this information to
-the email address specified in the <comment>--send-to</comment> option:
+                In addition to displaying the user list, you can also send this information to
+                the email address specified in the <comment>--send-to</comment> option:
 
-  <info>php %command.full_name%</info> <comment>--send-to=fabien@symfony.com</comment>
-
-HELP
+                  <info>php %command.full_name%</info> <comment>--send-to=fabien@symfony.com</comment>
+                HELP
             )
             // commands can optionally define arguments and/or options (mandatory and optional)
             // see https://symfony.com/doc/current/components/console/console_arguments.html
