@@ -15,7 +15,6 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Security\Core\Security;
 use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
 use Symfony\Component\Security\Http\Util\TargetPathTrait;
 
@@ -30,13 +29,11 @@ class SecurityController extends AbstractController
 {
     use TargetPathTrait;
 
-    /**
-     * @Route("/login", name="security_login")
-     */
-    public function login(Request $request, Security $security, AuthenticationUtils $helper): Response
+    #[Route('/login', name: 'security_login')]
+    public function login(Request $request, AuthenticationUtils $helper): Response
     {
         // if user is already logged in, don't display the login page again
-        if ($security->isGranted('ROLE_USER')) {
+        if ($this->getUser()) {
             return $this->redirectToRoute('blog_index');
         }
 
@@ -59,9 +56,8 @@ class SecurityController extends AbstractController
      *
      * But, this will never be executed. Symfony will intercept this first
      * and handle the logout automatically. See logout in config/packages/security.yaml
-     *
-     * @Route("/logout", name="security_logout")
      */
+    #[Route('/logout', name: 'security_logout')]
     public function logout(): void
     {
         throw new \Exception('This should never be reached!');
