@@ -24,9 +24,11 @@ use Twig\TwigFunction;
  */
 class AppExtension extends AbstractExtension
 {
-    private $localeCodes;
-    private $locales;
+    private array $localeCodes;
+    private ?array $locales = null;
 
+    // The $locales argument is injected thanks to the service container.
+    // See https://symfony.com/doc/current/service_container.html#binding-arguments-by-name-or-type
     public function __construct(string $locales)
     {
         $localeCodes = explode('|', $locales);
